If digilines is enabled, create digiline conducting copies of all registered pole nodes, with simple texture overlays to make it evident.
Also, create suitable spawners for these, but only if the light source
to be spawned is also digilines compatible.

When using a digiline-compat spawner, you can set the channel for the
light source (and thus all future lights you place) by shift-punching
while wielding the spawner.  This setting persists until the next
restart, or until you change it with another shift-punch.