Allow speciying custom top and overhang nodes
Can be true, false, or a table with two entries (topnode, overhang).

If false, the main pole extends all the way up to the specified height,
and the light is simply placed next to it.

if true, the main pole extends all the way up, and then steps right by
1m, with the light hanging from it.  This is the default.

if a table, the main pole stops 1m short of the top, and the two
specified nodes are placed above it, with the overhang to the right, and
the light hanging under it.

also fixed the custom base node not getting _digilines extension when
called-for